renamed eppic header and source file to world
