+++ src/core/world.h
